Initial client implementation
